separate out external only tests into their own file
