+++ source_gen/test/type_checker_test.dart
@@ -27,16 +27,13 @@ void main() {
   // Resolved top-level types from package:source_gen.
   late InterfaceType staticGenerator;
   late InterfaceType staticGeneratorForAnnotation;
-  late InterfaceType staticNonPublic;
   late TypeChecker staticGeneratorChecker;
   late TypeChecker staticGeneratorForAnnotationChecker;
-  late TypeChecker staticNonPublicChecker;
 
   setUpAll(() async {
     late LibraryElement core;
     late LibraryElement collection;
     late LibraryReader sourceGen;
-    late LibraryReader thisTest;
     await resolveSource(r'''
       export 'package:source_gen/source_gen.dart';
       export 'type_checker_test.dart' show NonPublic;
@@ -45,8 +42,6 @@ void main() {
       collection = (await resolver.findLibraryByName('dart.collection'))!;
       sourceGen = LibraryReader(await resolver
           .libraryFor(AssetId('source_gen', 'lib/source_gen.dart')));
-      thisTest = LibraryReader(await resolver
-          .libraryFor(AssetId('source_gen', 'test/type_checker_test.dart')));
     }, inputId: AssetId('source_gen', 'test/example.dart'));
 
     final staticIterable = core.getType('Iterable')!.instantiate(
@@ -85,9 +80,6 @@ void main() {
             nullabilitySuffix: NullabilitySuffix.none);
     staticGeneratorForAnnotationChecker =
         TypeChecker.fromStatic(staticGeneratorForAnnotation);
-    staticNonPublic = thisTest.findType('NonPublic')!.instantiate(
-        typeArguments: const [], nullabilitySuffix: NullabilitySuffix.none);
-    staticNonPublicChecker = TypeChecker.fromStatic(staticNonPublic);
   });
 
   // Run a common set of type comparison checks with various implementations.
@@ -97,7 +89,6 @@ void main() {
     required TypeChecker Function() checkHashMap,
     required TypeChecker Function() checkGenerator,
     required TypeChecker Function() checkGeneratorForAnnotation,
-    required TypeChecker Function() checkNonPublic,
   }) {
     group('(Iterable)', () {
       test('should be assignable from dart:collection#UnmodifiableListView',
@@ -183,19 +174,6 @@ void main() {
                 '${staticGeneratorForAnnotation.element.name}');
       });
     });
-
-    group('NonPublic', () {
-      test('should equal NonPublic', () {
-        expect(checkNonPublic().isExactlyType(staticNonPublic), isTrue,
-            reason: '${checkNonPublic()} != ${staticNonPublic.element.name}');
-      });
-
-      test('should be assignable from NonPublic', () {
-        expect(checkNonPublic().isAssignableFromType(staticNonPublic), isTrue,
-            reason: '${checkNonPublic()} is not assignable from '
-                '${staticNonPublic.element.name}');
-      });
-    });
   }
 
   group('TypeChecker.forRuntime', () {
@@ -206,7 +184,6 @@ void main() {
       checkGenerator: () => const TypeChecker.fromRuntime(Generator),
       checkGeneratorForAnnotation: () =>
           const TypeChecker.fromRuntime(GeneratorForAnnotation),
-      checkNonPublic: () => const TypeChecker.fromRuntime(NonPublic),
     );
   });
 
@@ -217,7 +194,6 @@ void main() {
       checkHashMap: () => staticHashMapChecker,
       checkGenerator: () => staticGeneratorChecker,
       checkGeneratorForAnnotation: () => staticGeneratorForAnnotationChecker,
-      checkNonPublic: () => staticNonPublicChecker,
     );
   });
 
@@ -230,8 +206,6 @@ void main() {
           'package:source_gen/src/generator.dart#Generator'),
       checkGeneratorForAnnotation: () => const TypeChecker.fromUrl(
           'package:source_gen/src/generator_for_annotation.dart#GeneratorForAnnotation'),
-      checkNonPublic: () => const TypeChecker.fromUrl(
-          'asset:source_gen/test/type_checker_test.dart#NonPublic'),
     );
   });
 
@@ -429,6 +403,3 @@ void main() {
 
 final throwsUnresolvedAnnotationException =
     throwsA(const TypeMatcher<UnresolvedAnnotationException>());
-
-// Used to check identity of non-public classes
-class NonPublic {}
